Split configuration and types into modules
